fingerprint_blood_group_detection/requirements.txt
# File: python_ml/requirements.txt

# Deep Learning Framework
tensorflow==2.15.0
keras==2.15.0

# Image Processing
opencv-python==4.8.1.78
Pillow==10.1.0

# Data Science Libraries
numpy==1.24.3
pandas==2.1.3
matplotlib==3.8.2
seaborn==0.13.0
scikit-learn==1.3.2

# Progress Bars
tqdm==4.66.1

# Serial Communication (for ESP32)
pyserial==3.5

# Server Framework (for inference)
flask==3.0.0
flask-cors==4.0.0

# Utilities
python-dotenv==1.0.0